pyproject.toml
[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"

[tool.poetry]
name = "dinohub-seedwork-be"
version = "0.1.5"
authors = [
  "Tuan Cau Rao <tuan.pt@orai.io>",
]
description = "A seedwork for DinoHub service"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [
    { include = "dino_seedwork_be" },
]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
toolz = "^0.11.2"
multipledispatch = "^0.6.0"
validators = "^0.19.0"
PyJWT = {extras = ["crypto"], version = "^2.4.0"}
SQLAlchemy = { extras = ["asyncio"], version = "^1.4.37" }
pytest-asyncio = "^0.18.3"
psycopg2 = "^2.9.3"
asyncpg = "^0.25.0"
greenlet = "^1.1.2"
python-dotenv = "^0.20.0"
python-dateutil = "^2.8.2"
importlib-resources = "^5.7.1"
pytz = "^2022.1"
fastapi = "0.85.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
jsonpickle = "^2.2.0"
httpx = "^0.23.0"
google-cloud-storage = "^2.4.0"
Pillow = "^9.2.0"
python-multipart = "^0.0.5"
returns = "^0.19.0"
pre-commit = "^2.20.0"
Faker = "^14.1.0"
respx = "^0.19.2"
pika = "^1.3.1"
multimethod = "^1.9"
firebase-admin = "^6.0.1"
sqlalchemy-utils = "^0.38.3"
pyramda = "^0.1"

[tool.poetry.dev-dependencies]
mypy = "^0.981"
coverage = "^6.4.4"
pytest = "7.2.0"
pytest-sugar = "0.9.6"
pytest-env = "^0.8.1"
pytest-instafail = "^0.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/mypy.toml"
